#3 mysql support for orm

+++ orm/src/test/java/rating_test/database/domain/entities/AuctionTable.kt
@@ -1,24 +1,23 @@
 package rating_test.database.domain.entities
 
-import rating_test.database.entities.User
 import ru.astrainteractive.astralibs.orm.database.Column
 import ru.astrainteractive.astralibs.orm.database.Constructable
 import ru.astrainteractive.astralibs.orm.database.Entity
 import ru.astrainteractive.astralibs.orm.database.Table
 
 
-object AuctionTable : Table<Long>("auctions") {
-    override val id: Column<Long> = long("id").primaryKey().autoIncrement()
+object AuctionTable : Table<Int>("auctions") {
+    override val id: Column<Int> = integer("id").primaryKey().autoIncrement()
     val discordId = text("discord_id").nullable()
     val minecraftUuid = text("minecraft_uuid")
-    val time = long("time")
+    val time = bigint("time")
     val item = byteArray("item")
     val price = float("price")
     val expired = bool("expired")
 }
 
 
-class Auction : Entity<Long>(AuctionTable) {
+class Auction : Entity<Int>(AuctionTable) {
     val id by AuctionTable.id
     val discordId by AuctionTable.discordId
     val minecraftUuid by AuctionTable.minecraftUuid
